pyproject.toml
[project]
name = "super-pocket"
version = "1.2.0"
description = "A collection of developer tools including markdown rendering, project exporters, agent templates, and cheatsheets management."
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Dimitri Gaggioli", email = "dimitri.gaggioli@gmail.com"}
]
keywords = ["dev-tools", "cli", "markdown", "productivity", "agents", "templates", "job", "assistant", "xml", "llm"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "click>=8.3.1",
    "rich-click>=1.9.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "httpx>=0.28.1",
    "uvicorn>=0.38.0",
    "fastapi>=0.121.3",
    "jinja2>=3.1.6",
    "pyyaml>=6.0.3",
    "pillow>=12.0.0"
]


[project.scripts]
pocket = "super_pocket.cli:cli"
markd = "super_pocket.markdown.renderer:markd"
proj2md = "super_pocket.project.to_file:proj_to_file"
req-to-date = "super_pocket.project.req_to_date:req_to_date_cli"
conv2pdf = "super_pocket.pdf.converter:main"
favicon = "super_pocket.web.favicon:main"
xml = "super_pocket.xml.cli:xml"
readme = "super_pocket.readme.cli:readme_cli"
project-init = "super_pocket.project.init.cli:new"
documents = "super_pocket.documents.cli:documents_cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=pocket --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["super_pocket/"]
omit = [
    "*/tests/*",
    "*/__init__.py"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.black]
line-length = 88
target-version = ["py311"]
